Update the docs
