+++ setup.py
@@ -42,7 +42,7 @@ setup(
     setup_requires=setup_requirements,
     test_suite='tests',
     tests_require=test_requirements,
-    url='https://github.com/emadmokhtar/pytafseer',
+    url='https://github.com/Quran-Tafseer/pytafseer/',
     version='0.1.0',
     zip_safe=False,
 )
